+++ asf_search/search/baseline_search.py
@@ -84,18 +84,21 @@ def get_stack_opts(
         raise ASFBaselineError(f'Requested reference product needs a baseline stack ID but does not have one: {reference.properties["fileID"]}')
 
     # build a stack from scratch if it's a non-precalc dataset with state vectors
-    if reference.properties['platform'].upper() in [PLATFORM.SENTINEL1A.upper(), PLATFORM.SENTINEL1B.upper()]:
+    
+    if reference.properties['processingLevel'] == 'BURST':
+        stack_opts.fullBurstID = reference.properties['burst']['fullBurstID']
+        stack_opts.polarization = [reference.properties['polarization']]
+        return stack_opts
+    elif reference.properties['platform'].upper() in [PLATFORM.SENTINEL1A.upper(), PLATFORM.SENTINEL1B.upper()]:
         stack_opts.platform = [PLATFORM.SENTINEL1A, PLATFORM.SENTINEL1B]
         
         stack_opts.beamMode = [reference.properties['beamModeType']]
         stack_opts.flightDirection = reference.properties['flightDirection']
         stack_opts.relativeOrbit = [int(reference.properties['pathNumber'])]  # path
         
-        is_burst = reference.properties['processingLevel'] == 'BURST'
-        
-        if reference.properties['polarization'] in ['HH', 'HH+HV'] and not is_burst:
+        if reference.properties['polarization'] in ['HH', 'HH+HV']:
             stack_opts.polarization = ['HH','HH+HV']
-        elif reference.properties['polarization'] in ['VV', 'VV+VH'] and not is_burst:
+        elif reference.properties['polarization'] in ['VV', 'VV+VH']:
             stack_opts.polarization = ['VV','VV+VH']
         else:
             stack_opts.polarization = [reference.properties['polarization']]